requirements.txt
# trace-processor/requirements.txt

# Core dependencies for pdf_to_text.py
google-cloud-aiplatform>=1.38.0
mistralai>=0.1.0
pinecone>=5.0.0
tqdm>=4.66.1

# Core dependencies for consumer.py
confluent-kafka==2.3.0

# Additional dependencies for both scripts
typing-extensions>=4.8.0  # For type hints support
google-cloud-storage>=2.7.0